requirements.txt
arrow==0.6.0
EbookLib==0.15
python-slugify==1.1.3
readability-lxml==0.6.1
requests==2.7.0
boto3==1.2.3
